# docs/swagger.yaml
basePath: /
definitions:
  functions.AiSearchPayload:
    properties:
      chatModel:
        $ref: '#/definitions/functions.ChatModel'
      focusMode:
        type: string
      optimizationMode:
        type: string
      query:
        type: string
      systemInstructions:
        type: string
    type: object
  functions.ChatModel:
    properties:
      customOpenAIBaseURL:
        type: string
      customOpenAIKey:
        type: string
      name:
        type: string
      provider:
        type: string
    type: object
  main.AiChatRequest:
    description: AI 聊天请求体
    properties:
      model_name:
        type: string
      query:
        type: string
    type: object
  main.AiChatResponse:
    description: AI 聊天响应体
    properties:
      response:
        type: string
    type: object
  main.AiSearchResponse:
    description: AI 搜索响应体
    properties:
      response:
        type: string
    type: object
  main.ChunithmAllSongResponse:
    description: CHUNITHM 全部歌曲响应体
    properties:
      songs:
        items: {}
        type: array
    type: object
  main.ChunithmSongResponse:
    description: CHUNITHM 单曲响应体
    properties:
      song: {}
    type: object
  main.CommonResponse:
    description: 通用响应体
    properties:
      message:
        type: string
      status:
        type: string
      version:
        type: string
    type: object
  main.MarkdownToPicRequest:
    description: Markdown转图片请求体
    properties:
      markdown:
        type: string
    type: object
  main.MarkdownToPicResponse:
    description: Markdown转图片响应体
    properties:
      url:
        type: string
    type: object
  main.StatusResponse:
    description: 状态响应体
    properties:
      status:
        type: string
    type: object
host: localhost:3000
info:
  contact:
    email: fiber@swagger.io
    name: API Support
  description: 这是一个基于 Fiber 的后端 API 示例，包含 AI、音游、Markdown 转图片等功能。
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: CongMiaoBot 后端 API
  version: "1.0"
paths:
  /:
    get:
      consumes:
      - application/json
      description: 获取后端运行状态
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.CommonResponse'
      summary: 主页状态
      tags:
      - 基础
  /ai/chat:
    post:
      consumes:
      - application/json
      description: 与 AI 进行对话
      parameters:
      - description: 请求体
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/main.AiChatRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.AiChatResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: AI 聊天
      tags:
      - AI
  /ai/search:
    post:
      consumes:
      - application/json
      description: 使用 AI 进行搜索
      parameters:
      - description: 请求体
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/functions.AiSearchPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.AiSearchResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: AI 搜索
      tags:
      - AI
  /functions/md-to-pic:
    post:
      consumes:
      - application/json
      description: 将 Markdown 内容渲染为图片
      parameters:
      - description: 请求体
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/main.MarkdownToPicRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.MarkdownToPicResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Markdown 转图片
      tags:
      - 工具
  /functions/random-pic/{folder}:
    get:
      description: 从指定文件夹随机返回一张图片
      parameters:
      - description: 文件夹名称
        in: path
        name: folder
        required: true
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: OK
          schema:
            type: file
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: 随机图片
      tags:
      - 工具
  /otoge/chunithm/allsong:
    get:
      consumes:
      - application/json
      description: 获取 CHUNITHM 所有歌曲信息
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.ChunithmAllSongResponse'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: 获取所有 CHUNITHM 歌曲
      tags:
      - 音游
  /otoge/chunithm/cover/{id}:
    get:
      description: 通过 ID 获取 CHUNITHM 歌曲封面图片
      parameters:
      - description: 歌曲ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: OK
          schema:
            type: file
      summary: 获取 CHUNITHM 歌曲封面
      tags:
      - 音游
  /otoge/chunithm/song/{id}:
    get:
      consumes:
      - application/json
      description: 通过 ID 获取 CHUNITHM 歌曲信息
      parameters:
      - description: 歌曲ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.ChunithmSongResponse'
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: 获取指定 ID 的 CHUNITHM 歌曲
      tags:
      - 音游
  /otoge/malody/info/{id}:
    get:
      consumes:
      - application/json
      description: 通过 ID 获取 Malody 用户信息
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: 获取 Malody 用户信息
      tags:
      - 音游
  /otoge/malody/recent/{id}:
    get:
      consumes:
      - application/json
      description: 通过 ID 获取 Malody 用户最近游玩记录
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: 获取 Malody 用户最近游玩记录
      tags:
      - 音游
  /status:
    get:
      consumes:
      - application/json
      description: 检查后端服务状态
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.StatusResponse'
      summary: 服务健康检查
      tags:
      - 基础
swagger: "2.0"